:recycle: TextButton 이용 리팩토링
